+++ src/Resources/app/administration/src/main.js
@@ -8,6 +8,8 @@ import './component/avalara-address-test-button';
 import './component/support-form';
 import './component/avalara-order-states';
 import './component/avalara-company-code-select';
+import './service/avalaraTaxCodeService';
+import './component/avalara-tax-code-select';
 
 // module for advanced custom settings
 import './module/mopt-avalara';